+++ libfwupd/fwupd-client.c
@@ -32,6 +32,7 @@
 
 #include "fwupd-client.h"
 #include "fwupd-common.h"
+#include "fwupd-deprecated.h"
 #include "fwupd-enums.h"
 #include "fwupd-error.h"
 #include "fwupd-device-private.h"
