+++ libfwupd/fwupd-result.c
@@ -26,6 +26,7 @@
 #include <string.h>
 
 #include "fwupd-common-private.h"
+#include "fwupd-deprecated.h"
 #include "fwupd-device-private.h"
 #include "fwupd-enums-private.h"
 #include "fwupd-error.h"
