+++ libfwupd/meson.build
@@ -16,6 +16,7 @@ install_headers(
 install_headers([
     'fwupd-client.h',
     'fwupd-common.h',
+    'fwupd-deprecated.h',
     'fwupd-device.h',
     'fwupd-enums.h',
     'fwupd-error.h',
